requirements.txt
streamlit>=1.40
chromadb>=0.5
sentence-transformers>=3.0
google-genai>=1.0.0
python-dotenv>=1.0
PyYAML>=6.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.2
pydantic>=2.7
tenacity>=9.0
pytest>=8.0